+++ src/app/examples/kai-table-example/kai-table-full-example.ts
@@ -7,6 +7,7 @@ import { IbKaiTableState } from "../../inobeta-ui/ui/kai-table/table.types";
   selector: "ib-kai-table-full-example",
   template: `
     <ib-kai-table
+      stripedRows
       tableName="fullExample"
       [displayedColumns]="columns"
       [data]="data"
@@ -64,7 +65,7 @@ import { IbKaiTableState } from "../../inobeta-ui/ui/kai-table/table.types";
         flex-direction: column;
         padding: 30px;
         gap: 3em;
-      }      
+      }
     `,
   ],
   providers: [UserService],
@@ -85,12 +86,10 @@ export class IbKaiTableFullExamplePage {
 
   getUserOrders() {
     this.state = "loading";
-    this.userService
-      .getUserOrders()
-      .subscribe((orders) => {
-        this.data = orders
-        this.state = "idle";
-      });
+    this.userService.getUserOrders().subscribe((orders) => {
+      this.data = orders;
+      this.state = "idle";
+    });
   }
 
   selectionChange(data) {